requirements.txt
Django==3.2.12
djangorestframework==3.14.0
PyJWT==1.7.1
PyJWT==2.3.0
PyJWT==2.8.0
pyotp==2.9.0
